+++ Sources/BraintreeCard/BTCardRequest.swift
@@ -1,6 +1,8 @@
 import Foundation
 
 /// Contains information about a card to tokenize
+// TODO: - NEXT_MAJOR_VERSION remove this class
+@available(*, deprecated, message: "Use BTCard directly instead")
 @objcMembers public class BTCardRequest: NSObject {
 
     /// The `BTCard` associated with this instance.