+++ CHANGELOG.md
@@ -7,6 +7,8 @@
   * Fix bug where defaults for `BTThreeDSecureRequest.accountType`, `BTThreeDSecureRequest.requestedExemptionType`, and `BTThreeDSecureRequest.dfReferenceID` were improperly returning an error if not passed into the request
 * BraintreeCard
   * Deprecate unused `BTCardRequest` class
+* BraintreeDataCollector
+  * Update previously incorrect version of PPRiskMagnes with 5.4.1 for Carthage users
 
 ## 5.24.0 (2023-10-30)
 * BraintreePayPalDataCollector